+++ src/legacy/main/java/org/lavro/carpetlir/mixins/FluidBlockMixin.java
@@ -2,10 +2,10 @@ package org.lavro.carpetlir.mixins;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.block.FluidBlock;
-import net.minecraft.registry.tag.FluidTags;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.lavro.carpetlir.features.renewable.CalciteGeneratorFeature;
+import org.lavro.carpetlir.helpers.FluidTagCompatibility;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -19,7 +19,7 @@ public abstract class FluidBlockMixin {
      */
     @Inject(method = "receiveNeighborFluids", at = @At("HEAD"), cancellable = true)
     private void carpetlir$handleCalciteBranch(World world, BlockPos pos, BlockState state, CallbackInfoReturnable<Boolean> cir) {
-        if (!state.getFluidState().isIn(FluidTags.LAVA)) {
+        if (!FluidTagCompatibility.isLava(state.getFluidState())) {
             return;
         }